cutscene dedup WIP
