+++ src/st/cutscene_dialog.h
@@ -1,0 +1,18 @@
+// maybe this can safely go into <cutscene.h>
+
+enum DialogueSteps {
+    DIALOGUE_INIT = 0,
+    DIALOGUE_RUN = 1,
+    DIALOGUE_LOAD_PORTRAIT = 2,
+    DIALOGUE_START_TEXT = 3,
+    DIALOGUE_UNLOAD_PORTRAIT = 4,
+    DIALOGUE_OPEN_DIALOG_BOX = 5,
+    DIALOGUE_CLOSE_DIALOG_BOX = 6,
+    DIALOGUE_END = 7,
+};
+
+enum DialogueSubsteps {
+    DIALOG_BOX_INIT = 0,
+    DIALOG_BOX_DRAW_RED = 1,
+    DIALOG_BOX_DRAW_BLUE = 2,
+};
