+++ src/st/dre/e_cutscene_dialogue.c
@@ -2,22 +2,7 @@
 #include "dre.h"
 #include <cutscene.h>
 
-enum DialogueSteps {
-    DIALOGUE_INIT = 0,
-    DIALOGUE_RUN = 1,
-    DIALOGUE_LOAD_PORTRAIT = 2,
-    DIALOGUE_START_TEXT = 3,
-    DIALOGUE_UNLOAD_PORTRAIT = 4,
-    DIALOGUE_OPEN_DIALOG_BOX = 5,
-    DIALOGUE_CLOSE_DIALOG_BOX = 6,
-    DIALOGUE_END = 7,
-};
-
-enum DialogueSubsteps {
-    DIALOG_BOX_INIT = 0,
-    DIALOG_BOX_DRAW_RED = 1,
-    DIALOG_BOX_DRAW_BLUE = 2,
-};
+#include "../cutscene_dialog.h"
 
 #define DAI_CUTSCENE_ALUCARD_READY 1
