+++ src/st/top/cutscene.c
@@ -200,13 +200,9 @@ void OVL_EXPORT(EntityCutscene)(Entity* self) {
                 self->step = 5;
                 self->step_s = 0;
                 return;
-            case CSOP_CLOSE_DIALOG:
-                if (g_SkipCutscene) {
-                    continue;
-                }
-                g_Dialogue.portraitAnimTimer = 0x18;
-                self->step = 6;
-                return;
+
+            #include "../cutscene_actions1.h"
+
             case CSOP_PLAY_SOUND:
                 if (g_SkipCutscene) {
                     g_Dialogue.scriptCur++;
